Update APIs
I changed the result type from 2 element tuple to 3 element tuple a
while ago but this module was commented out.. There must be a better way
to skip tests and still get compile checks.

Also, the old private aes was being called in problem7 so I updated it
ot the ecb specific version.