Sets up scaffolding for breaking ecb
Cryptopals problem 12: If you are given an oracle encryption function
that always uses the same unknown random key.. can you break it?

This commit sets up the scaffolding for preforming this attack (padding
oracle). So far I've set up the encryption oracle and determined the
length of the plaintext block (16) and written a function that spits out
a msg I can use as plaintext that will cause the PT being encrypted to
be an even number of blocks.

This sets up the ability to send in a block with exactly one byte two
few for a even block length. Instead of a 16 byte padding, you know that
the last byte of plaintext is 0x01 to indicate the single padding byte.
This can be leveraged to break the encryption totally.