Update problem specification
Initially I thought our text was appended... but actually the unknown
text is appended.

If we weren't told that our text was being prepended (and we weren't
writing the oracle function for practice ourselves..) we could verify
where our text was by submitting a text that causes the entire message
to be a byte too short (last byte will be padding byte of 1), then we
can submit the same message but with x01 appended to make the block size
even. If the second to last block (now there is a full padding block) is
the same we know our message was at the end, otherwise it's at the
beginning.