Mostly wip, breaks first block
Need to expand code to continue cracking remaining blocks.

This is a pretty neat attack so far. The first byte strategy is pretty
straightforward. submit 15 bytes and you know that the last byte of the
first block will be the first byte of the hidden message (the one you
don't know). Then submit 16 bytes built from the same 15 and another
byte (iterate through all u8s). You can then try to match the first
block of the new ciphertext with the old and when you get a match you
have the first byte of the hidden message.

For the second, you need to submit 14 bytes so that 2 bytes of the
hidden message are pulled into the first block. Then submit the same 14
bytes + the previously discovered plaintext byte + a random u8 (again
iterate through all u8's). When you get the same ciphertext back you've
found the second byte. Repeat until you have the entire first block.