Modify cracking function to recover all blocks
This involved some changes to modulo 16 the indexes to ensure I was
operating on the correct block as the while loop progressed. Getting the
main loop correct was a bit weird and the while condition may look
wrong.. and it would be wrong for certain plaintexts (those including
raw 0x01).

I believe a more accurate solution would likely continue to decrypt
until you couldn't find the next character.. This would indicate that
the character had changed when you submitted a different length oracle
message. I.e. you're at the end of the plaintext message and trying to
recover padding, which is changing as you adjust the length of input.
(i.e 0x01 -> 0x02 and now solving for the same block is impossible). I
will likely copy this code and update it for challenge 14 since it's a
related attack.