Adds profile struct
This allows decoding to an "object" (can't have two types in a hash,
though i could have just left uid as string..). Also, a nicer api for
encrypting profiles and decrypting blobs into profiles. This sets up the
scaffolding for the meat of challenge 13 which is to upgrade a user to
"admin" role using the encryption oracle.