Adds solution to challenge 13
This attack makes use of the profile encryption as an oracle. You can
submit any email address you want and get back the profile encoded and
encrypted. The encoding is as follows:

email=<email_address>&uid=<uid>&role=user

Because role is at the end and we are using ECB it's fairly easy to
isolate the 'user' part of the encoding into it's own block (since we
control the email). We can also submit an email that contains a block of
'admin' + the relevant 11 bytes of spacing. This means we can simply
splice the two ciphertexts together and have a valid decryption into the
form:

email=<email_address>&uid=<uid>&role=admin

**I've hard coded the uid to be 10 in this problem like the challenge
shows, however if it was random, we may need to run this attack several
times always assuming a different length say 1..10 until we get a valid
ciphertext.