Solve challenge 14
This test is SLOOOOWWW.. unclear if it's due to a poor method of
cracking or a really slow aes implementation. Tempted to rewrite aes to
use the chip procedures.

Anyways.. the way I solved this was to prepend exactly two blocks of the
same byte before my "oracle string". This meant that regardless of the
random bytes added at the front of the message I could find the two
identical blocks next to each other and know that the next block was my
attack string. This necessitated running the encryption function over
and over and over again until the block boundaries magically fell in the
right place. Assuming uniform randomness only 1 out of every 15 tries
would be useful for the attack, the other ciphertexts I threw away.

This means that for each letter I tried (~128 for each character of the
target text) I had to run aes ~8 times before getting a valid
ciphertext. The target text was about 140 in length. So roughly 143360
aes iterations to recover it. If those each took about 1 second it would
be 39 minutes.. it probably took around 20-30 minutes.