Use a bitflipping attack to enable admin privleges
This was surprisingly easy.. The shape of the plaintext (i.e. what was
prepended/appended). No special counting of the blocks to determine the
length of bogus bytes to lead the attack string with since the prepend
string was a round 2 blocks (32 bytes). Then we just needed to send 16
bytes to have a full block we could manipulate and not care about the
scrambled plaintext (since it was "user submitted" anyways).

The reason this works is the decryption algorithm xors the next
decrypted ciphertext with the previous ciphertext block to determine the
plaintext. Since it's already gone through the decryption algorithm at
this point, if you control the ciphertext block you can xor the next
block with whatever you want.. in my case to yeild illegal ';' and '='
characters and upgrade privleges to admin.

In general, this seemed way too easy..