Implement CTR mode encryption
This mode is really cool because encryption and decryption are the exact
same. Use AES or some other PRF to generate a stream of random bytes
that you can then xor with the plaintext. Wasn't sure how to design the
Api.. just provide a hood for ctr_prf? or provide decrypt_ctr and
encrypt_ctr even though they do the exact same thing. Went with latter
strategy for now..