Solve challenge 24
The first is to break encryption when a PRNG is being used as the stream
cipher. This is fairly trivial in this case since we also know that a 16
bit key is used (so there are only 2**16 different PRNG states it can be
starting from). You simply iterate through all of them, checking for
whether the last few bytes of the message match the output (since the
known plaintext is at the end). This would take a lot longer if the key
was 32 bits but certainly still doable.

The second part of the challenge was to determine whether a given
password reset token was generated using a PRNG with a unix timestamp
seed (this is apparently a thing..). You can do this by assuming the
timestamp was within the past few seconds (since you know exactly when
you receive the token) and try to match the first 32 bits of the token
with the first output from the PRNG seeded with a unix timestamp.