Break random access read/write
This is pretty trivial if the edit function lets you change anything.
Just edit the entire string with all a's and then xor the output with
the a's string. You then have the ctr stream and can xor the original
ciphertext to recover the plaintext. If they didn't allow such a large
edit.. you could still do it block by block or byte by byte but it would
take a bit longer and a bit more code