Break CBC using Key as IV
This was fairly trivial. The challenge set actually fully broke down one
method (implemented here as method2). I used another method that
involved zeroing the bits for the first two blocks, getting the
decrypted plaintext and then xor'ing the first two blocks. This works
because the second block that is returned is then the result of feeding
the zero block through AES. The first is the result of feeding through
the zero block xored with the IV (which is the key). If you xor the two
together the zero block "decryption" blocks cancel each other out and
you're left with the key.