Implement hash length extension hack
Challenge 29. Sha1 is an MD style hash. This means that given a hash and
knowledge of what was hashed (specifically the length) you can append
anything you want to create a hash that matches a hash of the
original_hashed_msg || glue_padding || appended_msg.

If this happens to be used as a MAC you have a problem since you don't
need to know the secret that was appended to the original message to
extend the msg and create another hash that looks like it was created
with the secret.