Add length extension attack for MD4
The only major changes between this and the Sha1 attack is the little
endianess of the calcs/length. In Sha1 this was bigendian (most
significant byte first) but in MD4 the spec explicitly uses little
endian. I had to change around the byte order of the length in bits that
is appended to the last block as well as change around the byte order of
the hash itself. The internal state of MD4 could be thought of as big
endian but the order is reversed upon finalization and hash output.