Solves Problem 31 and 32
These two problems touch on the subject of early exit string compares.
If the subroutine evaluating the validity of a signature exits upon
finding a byte that does not match the expected valid signature, the
timing the response will be different for various incorrect signatures.

This exploit involves sending a LOT of signatures and timing how long
the server takes to reject them. The longer the rejection takes the more
correct bytes the submitted forged mac had. You can then go about
enumerating all the possible bytes and determining which one takes
longest and assuming that was the correct byte for that position.

This problem has a artificial delay added in to make it a bit easier.
Without the delay this attack is extremely hard to pull off over a
network since network congestion/latency/route selection is all variable
and will add noise to the response time.