Fixes bug with small messages
If the original message was < 64 bytes, the entire forged "original
message" would be included in the padding computation. However, this
ALSO includes the space for the password.

This wasn't being accounted for since I was adding the null bytes for
password to the forged msg prior to appending the non-padding original
message blocks and then the padding blocks. Therefore, when the original
message was small the null bytes placeholder for the key/passphrase got
added in twice.